FIX: async model relationships
After upgrading ember-forecast-io we started receiving errors related to requests made for fetching `data_points`, which in fact are supposed to be embedded. According to EmbeddedRecordsMixin documentation `If a model is declared embedded for deserialization (embedded: 'always' or deserialize: 'records'), then do not use async: true`.  And { async: true } is the default option for relationships.